Move advice generation to other file
